Remove column header/row names from the data model
This was inspired by the work in #36.  It makes row_name and
column_header computed properties.  This implementation makes use of
hybrid_properties where possible, which will allow us to keep more of
the logic in the ORM layer.

This does remove the ability to query rows by name or columns by header
at the DB layer, but it shouldn't be a problem because our tables are
relatively small.

To make these properties mutable, I'm looking into using setters:

https://docs.sqlalchemy.org/en/latest/orm/extensions/hybrid.html#defining-setters