Move cached methods outside of classes
This is done due to an unfortunate limitation in dogpile... It cannot
use the class instance when hashing a function call.  This makes caching
any class methods or properties almost useless.  There may be a
metaclass magic way of handling this, but for now we just move the
cached methods into standalone functions.