Change batch label change endpoint to not do validation
This sets the label changes in the database, but does not return any
validation or table information.  The client must manually call either
the /csv/:id or /csv/:id/validation to get the new validation
information.

Note: The validation is computed lazily--the validation endpoints don't
mutate any state.  However, computing it once will cache some of the
results so that subsequent calls should be faster.