Update base docker image
I'm hoping this will fix the bug that keeps reoccuring on the production
instance:

    unable to load shared object '/usr/lib/opencpu/library/curl/libs/curl.so':
      /usr/lib/x86_64-linux-gnu/libssl.so.1.1: version `OPENSSL_1_1_1' not found (required by /usr/lib/x86_64-linux-gnu/libcurl.so.4)

    In call:
    dyn.load(file, DLLpath = DLLpath, ...)

This updates to R 3.6.1 which is incompatible with some of the
requirements for metaboanalyst.  Because we aren't using that package, I
just removed it.  I'm sure we could remove most of the packages we are
installing, but it isn't worth sorting through them at this point.