Weaken validator check
The original check for whether all labels are strings is too strong:
labels can be plain numbers as well. By making the check `!== 'object'`,
we allow for the label value to be any plain string or number value, but
not `null` or `undefined`, or any more complex type (i.e., class
instances or JavaScript objects).