Throw an error object in boxplot render
Currently when an error is thrown in this code, it is thrown as a data
object so sentry detects each time it happens as a unique issue.  I
don't know why there is an error being thrown here, or why the box plot
is even being drawn on page load.  This will at least prevent sentry
issue spam from continuing.