Fix BoxPlot render error
BoxPlots are rendered twice consectively, once when mounted and again
after the data is loaded. There is a 200ms D3 transition that is applied
during render. When the transition from the first render has not
completed before the second render starts, it is interrupted. D3 throws
a promise exception whenever a transition is interrupted. This can
safely be ignored.

There is also some spaghetti removal around triggering updates in
BoxPlots.