Fix race condition NPE
There is a race condition when loading the initial dataset using the
`LOAD_DATASET` action. Components will use the `loadDataset` mixin and
assume that the `dataset` in the store is non-null. `loadDataset`
triggers the `LOAD_DATASET` action, which asynchronously loads the
dataset. Normally this works fine, but during abnormally long loads or
unlucky scheduling `dataset` will be accessed before it is loaded,
resulting in an NPE in a random part of the app.

We already have an `INITIALIZE_DATASET` mutation, so the solution is to
simply have the `loadDataset` mixin invoke it before triggering the
`LOAD_DATASET` action. Mutations are synchronous, so `dataset` is now
guaranteed to be initialized.