Clean up axisPlot reactivity
Components which used `axisPlot` were relying on an awkward misuse of
computed properties to trigger a render immediately after rendering the
first time. I think the original motivation was that the graph SVG
element was not mounted until after the first render. However this is
trivially solved by calling the `update` function which updates the SVG
in a `mounted` hook.

- Remove the `reactiveUpdate` component from all plots which use it
- Fix a bug where putting "" into an integer field breaks the PCA plot
- Fix a path typo (`wilcoxon_volacno`)