Validate normalization errors with non-numeric data
The `weight-volume` normalization method in the Transform table uses a
sample column to transform the values in the rest of the data. If that
column contains non-numeric data, the transform will fail and throw a
server error.

The solution is to validate that the selected column is strictly numeric
before normalizing, and throwing a validation error if it is not.