Fix ScreePlot D3 errors
`ScreePlot.vue` contained several bugs that generated lots of error spam
in the console.

- Cutoff line coordinates were `NaN`. Cutoffs are vertical lines that
deliniate the 50%, 80%, and 90% points on the scree plot. When the
cutoff was before the first index (at index 0), the D3 map was breaking
and returning `NaN`. The fix is to not render when the index is 0.
- Transitions from one PCA component to more than one PCA component can
throw string parsing errors. This is because of some really weird string
parsing on SVG statements that is now removed.
- Setting the number of PCA components to greater than the number of
available components could cause errors. ScreePlot will now only render
up to the number of possible PCA components, and larger inputs are
ignored.