Fix Heatmap console errors
`Heatmap.vue` generated some errors during normal usage. These are now
fixed.

- When first loading the page, the default width/height (0) had the
margin and the dendrogram width/height subtracted from them, which
resulted in an initally negative width/height. This caused an SVG
rendering error.
- Deselecting all of the `Metabolite Filter` options or all of the
`Sample Filter` options resulted in a 400 error. This is expected, but
the error was handled incorrectly and locked up the UI; the only way to
get interactivity back was to reload the page.