Fix help icons not being right aligned
The help icons in the toolbar menus that display a tooltip when clicked
were overlapping with the radio button text rather than being right
aligned.

There is a bug introduced in Vuetify 1.3 and fixed in 2.1 that prevented
them from being alignable. There is a hack that needs to be removed when
the vuetify version is bumped.