Use TypeScript in LoadingsPlot
Following the example of BoxPlot, update LoadingsPlot to use Typescript.

There was an attempt to move the rendering from D3 into Vue, but that
was stymied by some strange event handling around mouseover tooltips.
For now LoadingsPlot is still using the same rendering code as always.

`useAxisPlot` now takes destructured arguments, and consumers can
specify if they want the X axis at the top or the bottom of the plot.