Convert PcaPage to use Typescript
Also convert the three components on the PcaPage, ScorePlot, ScreePlot,
and LoadingsPlot.

Using the composition API on PcaPage created a circular dependency:
```
PcaPage.js -> usePlotData.ts -> store/index.js -> analyses.js ->
PcaPage.js
```

The problem was that `analyses.js` provides a list of all available
analyses and all data required to describe them, including the
components that should be used to render them. `store/index.js` imported
`analyses.js` to dynamically determine what analyses needed plot data
stores, which had the side affect of importing all components used in
those analyses.

The solution was to remove the data required for the store from
`analyses.js` and move it to `store/index.js`. The downside is that
analysis data is now defined in two places, but it seems appropriate
that visualization info is kept in `vis/analyses.js` while store info is
kept in the store.