Stabilize test navbar usage
Using the Analysis methods in the nav bar involves clicking the
`Analysis` nav bar option, which triggers an animation which reveals the
analyses. Tests were trying to click the analysis before it was
guaranteed to be revealed, which was causing flake.

Move opening an analysis page into a utility function.