Convert Pretreatment
This adds a circular dependency between `routes/index.js` and
`Pretreatment.vue`. `Pretreatment` is required in `routes` because it is
navigable in one of the default routes, while `routes` is required in
`Pretreatment` because it needs to trigger navigation to different
components. This was not a problem because `$routes` was injected into
2.0 component definitions, but this is no longer possible with
composition API.